- created new AsmTasksResource Class to limit the amount of code repetition inside the ASM classes - removed resulting reduntant code - all tests passing as expected.
