+++ f5/bigip/resource.py
@@ -1055,10 +1055,7 @@ class AsmResource(Resource):
     ASM resources are unique in BIG-IP® in the sense that their direct URI
     endpoints are hash based IDs of the resources.
 
-    The IDs are generated by BIG-IP® when the objects are created. Some of
-    the resources do not support create() method in the strict sense,
-    as they require an HTTP POST with an empty json{} to prompt BIGIP to
-    create them, therefore a new method fetch() was created.
+    The IDs are generated by BIG-IP® when the objects are created.
 
     Moreover, the ASM resources do not have 'generation' property,
     therefore some of the other methods needed to be adjusted to accommodate
@@ -1068,8 +1065,7 @@ class AsmResource(Resource):
     def __init__(self, container):
         """Call to create a client side object to represent a service URI.
 
-        Call _create _load or _fetch for a Resource resource to have a
-        self._meta_data['uri']!
+        Call _create or _load for a Resource to have a self._meta_data['uri']!
         """
         super(AsmResource, self).__init__(container)
         # Asm endpoints require object 'id' which is a hash created by BIGIP
@@ -1077,31 +1073,6 @@ class AsmResource(Resource):
         self._meta_data['required_load_parameters'] = set(('id',))
         # No ASM endpoint supports Stats
         self._meta_data['object_has_stats'] = False
-
-    def _fetch(self):
-        """wrapped by `fetch` override that in subclasses to customize"""
-        if 'uri' in self._meta_data:
-            error = "There was an attempt to assign a new uri to this "\
-                    "resource, the _meta_data['uri'] is %s and it should"\
-                    " not be changed." % (self._meta_data['uri'])
-            raise URICreationCollision(error)
-        # Make convenience variable with short names for this method.
-        _create_uri = self._meta_data['container']._meta_data['uri']
-        session = self._meta_data['bigip']._meta_data['icr_session']
-        # Invoke the REST operation on the device.
-        response = session.post(_create_uri, json={})
-        # Make new instance of self
-        return self._produce_instance(response)
-
-    def fetch(self):
-        """Fetch the ASM resource on the BIG-IP®.
-
-        This is a heavily modified version of create, that does not allow
-        any arguments when executing. It uses an emtpy json{} HTTP POST to
-        prompt the BIG-IP® to create the object, mainly used by 'Tasks'
-        endpoint.
-        """
-        return self._fetch()
 
     def _load(self, **kwargs):
         """wrapped with load, override that in a subclass to customize"""
@@ -1219,4 +1190,66 @@ class AsmResource(Resource):
         """
         raise UnsupportedOperation(
             "%s does not support the update method" % self.__class__.__name__
+        )
+
+
+class AsmTaskResource(AsmResource):
+    """ASM Task Resource class represents an ASM Tasks endpoint on the
+
+    device.
+
+    Tasks resources do not support create() method in the strict sense,
+    as they require an HTTP POST with an empty json{} to prompt BIGIP to
+    create them, therefore a new method fetch() was created.
+
+    """
+
+    def __init__(self, container):
+        """Call to create a client side object to represent a service URI.
+
+        Call  _fetch for a Resource to have a self._meta_data['uri']!
+        """
+        super(AsmTaskResource, self).__init__(container)
+
+    def _fetch(self):
+        """wrapped by `fetch` override that in subclasses to customize"""
+        if 'uri' in self._meta_data:
+            error = "There was an attempt to assign a new uri to this "\
+                    "resource, the _meta_data['uri'] is %s and it should"\
+                    " not be changed." % (self._meta_data['uri'])
+            raise URICreationCollision(error)
+        # Make convenience variable with short names for this method.
+        _create_uri = self._meta_data['container']._meta_data['uri']
+        session = self._meta_data['bigip']._meta_data['icr_session']
+        # Invoke the REST operation on the device.
+        response = session.post(_create_uri, json={})
+        # Make new instance of self
+        return self._produce_instance(response)
+
+    def fetch(self):
+        """Fetch the ASM resource on the BIG-IP®.
+
+        This is a heavily modified version of create, that does not allow
+        any arguments when executing. It uses an emtpy json{} HTTP POST to
+        prompt the BIG-IP® to create the object, mainly used by 'Tasks'
+        endpoint.
+        """
+        return self._fetch()
+
+    def create(self, **kwargs):
+        """Create is not supported for Task ASM resources
+
+                :raises: UnsupportedOperation
+        """
+        raise UnsupportedOperation(
+            "%s does not support the create method" % self.__class__.__name__
+        )
+
+    def modify(self, **kwargs):
+        """Modify is not supported for Task ASM resources
+
+                :raises: UnsupportedOperation
+        """
+        raise UnsupportedOperation(
+            "%s does not support the modify method" % self.__class__.__name__
         )
